prepare footer for storybook
