+++ src/components/footer/footer.story.js
@@ -1,0 +1,9 @@
+import React from "react"
+import { Footer } from "../"
+
+export default {
+  title: "Components|Footer",
+  component: Footer,
+}
+
+export const base = () => <Footer />
